+++ symbolic/src/Data/Macaw/Symbolic/Testing.hs
@@ -34,6 +34,7 @@ module Data.Macaw.Symbolic.Testing (
 
 import qualified Control.Exception as X
 import qualified Control.Lens as L
+import           Control.Monad ( when )
 import qualified Data.BitVector.Sized as BVS
 import qualified Data.ByteString as BS
 import qualified Data.ElfEdit as EE
@@ -75,6 +76,7 @@ import qualified What4.BaseTypes as WT
 import qualified What4.Expr as WE
 import qualified What4.FunctionName as WF
 import qualified What4.Interface as WI
+import qualified What4.LabeledPred as WL
 import qualified What4.ProgramLoc as WPL
 import qualified What4.Protocol.Online as WPO
 import qualified What4.SatResult as WSR
@@ -180,6 +182,42 @@ functionName dfi = maybe addrName fromByteString (MD.discoveredFunSymbol dfi)
     addrName = WF.functionNameFromText (Text.pack (show (MD.discoveredFunAddr dfi)))
     fromByteString = WF.functionNameFromText . Text.decodeUtf8With Text.lenientDecode
 
+proveOneGoal
+  :: ( CB.IsSymInterface sym
+     , sym ~ WE.ExprBuilder t st fs
+     )
+  => WS.SolverAdapter st
+  -> sym
+  -> CB.Assumptions sym
+  -> WL.LabeledPred (WI.Pred sym) CS.SimError
+  -> IO ()
+proveOneGoal goalSolver sym asmps lp = do
+  assumptions <- CB.assumptionsPred sym asmps
+  goal <- WI.notPred sym (lp L.^. WL.labeledPred)
+  WS.solver_adapter_check_sat goalSolver sym WS.defaultLogData [assumptions, goal] $ \satRes ->
+    case satRes of
+      WSR.Unsat {} -> return ()
+      WSR.Sat {} -> error ("Failed to prove goal: " ++ show (lp L.^. WL.labeledPredMsg))
+      WSR.Unknown {} -> error ("Failed to prove goal: " ++ show (lp L.^. WL.labeledPredMsg))
+  return ()
+
+proveGoals
+  :: ( CB.IsSymInterface sym, sym ~ WE.ExprBuilder t st fs )
+  => WS.SolverAdapter st
+  -> sym
+  -> Maybe (CB.Goals (CB.Assumptions sym) (CB.Assertion sym))
+  -> IO ()
+proveGoals goalSolver sym = mapM_ (go mempty)
+  where
+    go asmps gs =
+      case gs of
+        CB.Assuming as gs1 -> go (asmps <> as) gs1
+        CB.Prove lp -> proveOneGoal goalSolver sym asmps lp
+        CB.ProveConj g1 g2 -> do
+          go asmps g1
+          go asmps g2
+          return ()
+
 -- | Convert the given function into a Crucible CFG, symbolically execute it,
 -- and treat the return value as an assertion to be verified.
 --
@@ -198,8 +236,15 @@ functionName dfi = maybe addrName fromByteString (MD.discoveredFunSymbol dfi)
 -- *translation* of programs into Crucible, while 'MS.MacawArchEvalFn' is used
 -- during symbolic execution (to provide interpretations to
 -- architecture-specific primitives).
-simulateAndVerify :: forall arch sym ids w t st fs
-                   . ( CB.IsSymInterface sym
+--
+-- In addition to symbolically executing the function to produce a Sat/Unsat
+-- result, this function will attempt to verify all generate side conditions if
+-- the name of the function being simulated begins with @test_and_verify_@ (as
+-- opposed to just @test@).  This is necessary for testing aspects of the
+-- semantics that involve the generated side conditions, rather than just the
+-- final result.
+simulateAndVerify :: forall arch sym bak ids w t st fs
+                   . ( CB.IsSymBackend sym bak
                      , CCE.IsSyntaxExtension (MS.MacawExt arch)
                      , MM.MemWidth (MC.ArchAddrWidth arch)
                      , w ~ MC.ArchAddrWidth arch
@@ -212,7 +257,7 @@ simulateAndVerify :: forall arch sym ids w t st fs
                   -- ^ The solver adapter to use to discharge assertions
                   -> WS.LogData
                   -- ^ A logger to (optionally) record solver interaction (for the goal solver)
-                  -> sym
+                  -> bak
                   -- ^ The symbolic backend
                   -> [CS.GenericExecutionFeature sym]
                   -- ^ Crucible execution features, see 'defaultExecFeatures' for a good initial selection
@@ -227,7 +272,8 @@ simulateAndVerify :: forall arch sym ids w t st fs
                   -> MD.DiscoveryFunInfo arch ids
                   -- ^ The function to simulate and verify
                   -> IO SimulationResult
-simulateAndVerify goalSolver logger sym execFeatures archInfo archVals mem (ResultExtractor withResult) dfi =
+simulateAndVerify goalSolver logger bak execFeatures archInfo archVals mem (ResultExtractor withResult) dfi =
+  let sym = CB.backendGetSym bak in
   MS.withArchConstraints archVals $ do
     let funName = functionName dfi
     halloc <- CFH.newHandleAllocator
@@ -235,9 +281,9 @@ simulateAndVerify goalSolver logger sym execFeatures archInfo archVals mem (Resu
 
     let endianness = MSM.toCrucibleEndian (MAI.archEndianness archInfo)
     let ?recordLLVMAnnotation = \_ _ _ -> return ()
-    (initMem, memPtrTbl) <- MSM.newGlobalMemory (Proxy @arch) sym endianness MSM.ConcreteMutable mem
+    (initMem, memPtrTbl) <- MSM.newGlobalMemory (Proxy @arch) bak endianness MSM.ConcreteMutable mem
     let globalMap = MSM.mapRegionPointers memPtrTbl
-    (memVar, stackPointer, execResult) <- simulateFunction sym execFeatures archVals halloc initMem globalMap g
+    (memVar, stackPointer, execResult) <- simulateFunction bak execFeatures archVals halloc initMem globalMap g
     case execResult of
       CS.TimeoutResult {} -> return SimulationTimeout
       CS.AbortedResult {} -> return SimulationAborted
@@ -245,6 +291,10 @@ simulateAndVerify goalSolver logger sym execFeatures archInfo archVals mem (Resu
         case partialRes of
           CS.PartialRes {} -> return SimulationPartial
           CS.TotalRes gp -> do
+            when ("test_and_verify_" `Text.isPrefixOf` WF.functionName funName) $ do
+              goals <- CB.getProofObligations bak
+              proveGoals goalSolver sym goals
+
             let Just postMem = CSG.lookupGlobal memVar (gp L.^. CS.gpGlobals)
             withResult (gp L.^. CS.gpValue) stackPointer postMem $ \resRepr val -> do
               -- The function is assumed to return a value that is intended to be
@@ -252,7 +302,7 @@ simulateAndVerify goalSolver logger sym execFeatures archInfo archVals mem (Resu
               --
               -- The result is treated as true if it is not equal to zero
               zero <- WI.bvLit sym resRepr (BVS.mkBV resRepr 0)
-              bv_val <- CLM.projectLLVM_bv sym val
+              bv_val <- CLM.projectLLVM_bv bak val
               notZero <- WI.bvNe sym bv_val zero
               goal <- WI.notPred sym notZero
               WS.solver_adapter_check_sat goalSolver sym logger [goal] $ \satRes ->
@@ -278,14 +328,14 @@ simulateAndVerify goalSolver logger sym execFeatures archInfo archVals mem (Resu
 -- immutable contents of the text segment).
 simulateFunction :: ( ext ~ MS.MacawExt arch
                     , CCE.IsSyntaxExtension ext
-                    , CB.IsSymInterface sym
+                    , CB.IsSymBackend sym bak
                     , CLM.HasLLVMAnn sym
                     , MS.SymArchConstraints arch
                     , w ~ MC.ArchAddrWidth arch
                     , 16 <= w
                     , ?memOpts :: CLM.MemOptions
                     )
-                 => sym
+                 => bak
                  -> [CS.GenericExecutionFeature sym]
                  -> MS.ArchVals arch
                  -> CFH.HandleAllocator
@@ -296,7 +346,8 @@ simulateFunction :: ( ext ~ MS.MacawExt arch
                        , CLM.LLVMPtr sym w
                        , CS.ExecResult (MS.MacawSimulatorState sym) sym ext (CS.RegEntry sym (MS.ArchRegStruct arch))
                        )
-simulateFunction sym execFeatures archVals halloc initMem globalMap g = do
+simulateFunction bak execFeatures archVals halloc initMem globalMap g = do
+  let sym = CB.backendGetSym bak
   let symArchFns = MS.archFunctions archVals
   let crucRegTypes = MS.crucArchRegTypes symArchFns
   let regsRepr = CT.StructRepr crucRegTypes
@@ -313,8 +364,8 @@ simulateFunction sym execFeatures archVals halloc initMem globalMap g = do
   stackArrayStorage <- WI.freshConstant sym (WSym.safeSymbol "stack_array") WI.knownRepr
   stackSize <- WI.bvLit sym WI.knownRepr (BVS.mkBV WI.knownRepr (2 * 1024 * 1024))
   let ?ptrWidth = WI.knownRepr
-  (stackBasePtr, mem1) <- CLM.doMalloc sym CLM.StackAlloc CLM.Mutable "stack_alloc" initMem stackSize CLD.noAlignment
-  mem2 <- CLM.doArrayStore sym mem1 stackBasePtr CLD.noAlignment stackArrayStorage stackSize
+  (stackBasePtr, mem1) <- CLM.doMalloc bak CLM.StackAlloc CLM.Mutable "stack_alloc" initMem stackSize CLD.noAlignment
+  mem2 <- CLM.doArrayStore bak mem1 stackBasePtr CLD.noAlignment stackArrayStorage stackSize
 
   -- Make initial registers, including setting up a stack pointer (which points
   -- into the middle of the stack allocation, to allow accesses into the caller
@@ -333,7 +384,7 @@ simulateFunction sym execFeatures archVals halloc initMem globalMap g = do
   let fnBindings = CFH.insertHandleMap (CCC.cfgHandle g) (CS.UseCFG g (CAP.postdomInfo g)) CFH.emptyHandleMap
   MS.withArchEval archVals sym $ \archEvalFn -> do
     let extImpl = MS.macawExtensions archEvalFn memVar globalMap lookupFunction lookupSyscall validityCheck
-    let ctx = CS.initSimContext sym CLI.llvmIntrinsicTypes halloc IO.stdout (CS.FnBindings fnBindings) extImpl MS.MacawSimulatorState
+    let ctx = CS.initSimContext bak CLI.llvmIntrinsicTypes halloc IO.stdout (CS.FnBindings fnBindings) extImpl MS.MacawSimulatorState
     let s0 = CS.InitialState ctx initGlobals CS.defaultAbortHandler regsRepr simAction
     res <- CS.executeCrucible (fmap CS.genericToExecutionFeature execFeatures) s0
     return (memVar, sp, res)
@@ -343,10 +394,10 @@ simulateFunction sym execFeatures archVals halloc initMem globalMap g = do
 --
 -- The online constraints allow us to turn on path satisfiability checking
 data SomeBackend sym where
-  SomeOnlineBackend :: ( sym ~ CBO.OnlineBackend scope solver fs
+  SomeOnlineBackend :: ( sym ~ WE.ExprBuilder scope st fs
                        , WPO.OnlineSolver solver
                        , CB.IsSymInterface sym
-                       ) => sym -> SomeBackend sym
+                       ) => CBO.OnlineBackend solver scope st fs -> SomeBackend sym
   SomeOfflineBackend :: sym -> SomeBackend sym
 
 -- | A default set of execution features that are flexible enough to support a
@@ -357,8 +408,9 @@ defaultExecFeatures :: SomeBackend sym -> IO [CS.GenericExecutionFeature sym]
 defaultExecFeatures backend =
   case backend of
     SomeOfflineBackend {} -> return []
-    SomeOnlineBackend sym -> do
-      pathSat <- CSP.pathSatisfiabilityFeature sym (CBO.considerSatisfiability sym)
+    SomeOnlineBackend bak -> do
+      let sym = CB.backendGetSym bak
+      pathSat <- CSP.pathSatisfiabilityFeature sym (CBO.considerSatisfiability bak)
       return [pathSat]
 
 -- | This type wraps up a function that takes the post-state of a symbolic